+++ BloomFramework/include/Systems/RenderSystem.h
@@ -12,7 +12,7 @@ namespace bloom::systems {
 		using System::DefaultSystem;
 
 	public:
-		void update(uint64_t = 0) override {
+		void update(double = 0) override {
 			m_registry.view<Position, Size, Sprite>().each(
 				[](auto, Position& position, Size& size, Sprite& sprite) {
 					sprite.texture->render(sprite.srcRect, SDL_Rect{ position.x, position.y, size.w, size.h });
